fix: drop image.decode() from pdf_flatten.js to fix Linux headless hang
image.decode() neither resolves nor rejects in Linux headless Chrome
under --virtual-time-budget, so the box-shadow flatten silently stalled
and PDFs exported on Linux kept /S /Luminosity soft masks (the #150
black rectangles in Preview.app). Unify applyOuterShadow on the
existing loadImage load-event wait, and add a Linux e2e CI job
(explicit PEITHO_CHROME_PATH that panics when missing, so the E2E
suite can never silently skip in CI) to keep this class of bug from
shipping undetected again.

The sibling decode() usage in measure.ts was fixed here too originally,
but PPTX export (and measure.ts with it) was removed upstream by #156,
so the final diff covers pdf_flatten.js only.

Closes #155
