+++ CLAUDE.md
@@ -51,6 +51,7 @@ Always verify UX changes end-to-end in a real browser/real display (jsdom cannot
 - **Flag handoff to an already-running Chrome instance only works via `--app`**: `--start-fullscreen`/`--window-position`/`--window-size` are ignored. To reliably apply them, launch a new process with a separate `--user-data-dir` (which is why slides/presenter use two instances, `~/.peitho/chrome-profile-{slides,presenter}`)
 - **PDF export flattens gradients at print time**: `pdf.html` embeds a script that converts CSS gradients to Chrome-rasterized images because Type 4 shadings render pathologically slowly and incorrectly in macOS Preview.app (measured 2026-07-06). Design record: `docs/plans/2026-07-06-pdf-gradient-flatten.md`
 - **PDF export flattens box-shadow at print time**: Chrome emits blurred CSS box-shadows as `/S /Luminosity` soft masks that Quartz renders as hard black rectangles, so `pdf_flatten.js` rasterizes supported shadows to RGBA PNGs for PDF export (measured 2026-07-06). Design record: `docs/plans/2026-07-06-pdf-shadow-flatten.md`
+- **`image.decode()` hangs forever (neither resolves nor rejects) in Linux headless Chrome under `--virtual-time-budget`** (measured 2026-07-06, Issue #155): no rejection means `.catch()` never fires — the hang is completely silent, and Chrome prints the page once the virtual-time budget expires. On macOS decode just happened to finish before the budget ran out. In-page scripts run under virtual time (`pdf_flatten.js`) therefore never rely on `decode()` and wait on load/error events instead; reintroduction is guarded by the Linux `e2e` CI job (with `PEITHO_CHROME_PATH` set explicitly so a missing Chrome panics instead of silently skipping). Design record: `docs/plans/2026-07-06-pdf-flatten-linux-decode-hang.md`
 - **On macOS, Chrome's process lingers even after all windows are closed**: if the previous present instance is still holding the peitho profile, the next launch becomes a handoff and all placement flags are lost. That's why present terminates any lingering process before opening
 - **Lingering Chrome processes must be terminated normally, not killed with SIGTERM**: SIGTERM registers as a crash to Chrome (`exit_type: Crashed`), triggering crash recovery on the next launch which restores the old session's windows/bounds. Using `NSRunningApplication.terminate` (in JXA this fires via **parenthesis-less** access, due to the bridge — confirmed by testing) terminates it as Normal. Note that `exit_type` always shows Crashed while running (it's written up front at launch, then flipped back to Normal on clean exit), so reading it while the process is active is meaningless. Restrict target pids to the Chrome main process (no `--type=`) from `ps` (`pgrep -f` also catches the shell itself if it contains the pattern)
 - **Chrome 149 on macOS may never exit after one-shot headless PDF export**: `--print-to-pdf` can complete its output but keep the parent process alive after waking GoogleUpdater. Export code must use the shared one-shot runner (spawn + piped readers + completion predicate + timeout + kill/reap), never `Command::output()`. Killing is acceptable here because export uses throwaway temp `--user-data-dir` profiles; the SIGTERM crash-recovery warning above applies to persistent presenter profiles
